api/requirements.txt
requests
requests-toolbelt
google-cloud-datastore==2.4.0
Flask==1.1.2
Flask-RESTful==0.3.8
PyJWT==1.4.0
Pyrebase
google-cloud==0.34.0
google-cloud-firestore==1.9.0
backports-datetime-fromisoformat==1.0.0
certifi==2021.5.30
charset-normalizer==2.0.6
coverage==6.0
gitdb==4.0.7
GitPython==3.1.0
idna>=2.5
python-dotenv==0.19.0
smmap==4.0.0
typing-extensions==3.10.0.2
urllib3<1.26,>=1.21.1
